backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.3.1
python-decouple==3.8
psycopg2-binary==2.9.9
pandas==2.1.3
openpyxl==3.11.0
gunicorn==21.2.0
